fix issues with unicode characters
dont fail on a bad file
